Implement graceful shutdown.
* Call acpi::poweroff() from arch::x86_64::processor::shutdown().

* Store the exit code of the last task that finished.

* Add a sys_shutdown() syscall implemented per SysCallInterface.
  The generic implementation just calls arch::processor::shutdown(), whereas the Proxy and Uhyve SyscallInterfaces send SysExit
  commands over their communication channels.
  This supersedes the per-SyscallInterface sys_exit() call, which would wrongly shut down the entire system when the first task exited.

* Add a sys_lwip_register_tcpip_task() syscall to register the Task ID of the lwIP TCP/IP thread.
  This enables us to shut down the system when only the lwIP TCP/IP thread is left.

* Make TaskId u32 to be consistent with the Tid type of the syscalls.

+++ librs/src/scheduler/task.rs
@@ -56,14 +56,14 @@ pub enum WakeupReason {
 
 /// Unique identifier for a task (i.e. `pid`).
 #[derive(PartialEq, Eq, PartialOrd, Ord, Debug, Clone, Copy)]
-pub struct TaskId(usize);
+pub struct TaskId(u32);
 
 impl TaskId {
-	pub const fn into(self) -> usize {
+	pub const fn into(self) -> u32 {
 		self.0
 	}
 
-	pub const fn from(x: usize) -> Self {
+	pub const fn from(x: u32) -> Self {
 		TaskId(x)
 	}
 }
